requirements.txt
fastapi
pydantic
uvicorn
watchfiles
gunicorn
starlette
bs4
requests
python-multipart
python-jose[cryptography]
passlib[bcrypt]